refactor: separate action buttons into components
